+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [vtex test e2e] Added retry count to account for cold `tester-hub` start
+- [publish/link] Ignore Cypress generated videos/screenshots
 
 ## [2.101.0] - 2020-05-19
 ### Added