+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [vtex test e2e] Added retry count to account for cold `tester-hub` start
-- [publish/link] Ignore Cypress generated videos/screenshots
+
+### Changed
+- [vtex publish/link] Ignore Cypress generated videos/screenshots
 
 ## [2.101.0] - 2020-05-19
 ### Added