+++ CHANGELOG.md
@@ -5,11 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixed
-- [vtex login, vtex switch] Prompt workspace creation if it doesn't exist.
+### Added
+- [vtex test e2e] Added retry count to account for cold `tester-hub` start.
 
 ### Changed
+- [vtex publish/link] Ignore Cypress generated videos/screenshots.
 - [ErrorReport] Do not send wrong usage errors to telemetry.
+
+### Fixed
+- [vtex login, vtex switch] Prompt workspace creation if it doesn't exist.
 
 ## [2.101.0] - 2020-05-19
 ### Added