+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [vtex test e2e] Added retry count to account for cold `tester-hub` start.
+- [vtex link] Generate `linkID` and set the `x-vtex-bh-link-id` header on link and relink requests to `builder-hub`.
 
 ### Changed
 - [vtex publish/link] Ignore Cypress generated videos/screenshots.
@@ -14,9 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - [vtex login, vtex switch] Prompt workspace creation if it doesn't exist.
-
-### Added
-- [vtex link] Generate `linkID` and set the `x-vtex-bh-link-id` header on link and relink requests to `builder-hub`.
 
 ## [2.101.0] - 2020-05-19
 ### Added